Split encoder/decoder into specialized text or stream encoder/decoders.
This change was made to support the use of text input streams that
need to be read in different character encodings. Since the inputStream
is encoding-agnostic, we cannot simply pipe in a string of characters
and assume it all to work. Thus, we've now added a different set of
interfaces to support character readers and writers, and adjusted the
interfaces and tests accordingly. Unfortunately, this also meant that our
dynamically-detecting encoders/decoders can no longer provide a single
interface, so they had to be removed.